Some refactoring for pr
