+++ build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'org.jetbrains.kotlin.jvm' version '1.2.51'
 }
 
-group 'completion-plugin'
+group 'com.intellij.completion.evaluation'
 version '1.0-SNAPSHOT'
 
 repositories {
