+++ settings.gradle
@@ -1,2 +1,1 @@
-rootProject.name = 'completion-plugin'
-
+rootProject.name = 'completion-evaluation-plugin'
