+++ src/main/resources/META-INF/plugin.xml
@@ -1,5 +1,5 @@
 <idea-plugin>
-    <id>completion-plugin.completion-plugin</id>
+    <id>groupId.artifactId</id>
     <name>Plugin display name here</name>
     <vendor email="support@yourcompany.com" url="http://www.yourcompany.com">YourCompany</vendor>
 
@@ -20,7 +20,7 @@
 
     <actions>
         <!-- Add your actions here -->
-        <action id="PrintFileAction" class="PrintFileAction" text="Print File"
+        <action id="org.jb.cce.actions.PrintFileAction" class="org.jb.cce.actions.GetClassPSIAction" text="Print File"
                 description="Prints current file into console">
             <add-to-group group-id="EditorPopupMenu" anchor="first"/>
         </action>
